// src/analysis/types.ts
export interface JavaScriptEntity extends BaseCodeEntity {
  type:
    | "function"
    | "class"
    | "interface"
    | "type"
    | "import"
    | "export"
    | "variable";
  isAsync?: boolean;
  isExported?: boolean;
  isAbstract?: boolean;
  isDefault?: boolean;
  isReExport?: boolean;
  module?: string;
  exportType?: "named" | "default" | "all" | "namespace" | "assignment";
}

export interface RustEntity extends BaseCodeEntity {
  type:
    | "function"
    | "struct"
    | "enum"
    | "trait"
    | "implementation"
    | "module"
    | "import"
    | "constant"
    | "static"
    | "type-alias";
  language: "rust";
}

export type CodeEntity = JavaScriptEntity | RustEntity;

export interface CallExpression {
  callee: string;
  type: "function" | "method" | "dynamic" | "unknown";
  line: number;
  containingFunction: string | null;
  signature: string;
  argumentCount: number;
}

export interface TypeAlias {
  name: string;
  line: number;
  typeParameters: string[];
  definition: string;
  isExported: boolean;
}

export interface InterfaceInfo {
  name: string;
  line: number;
  typeParameters: string[];
  extends: string[];
  properties: number;
  isExported: boolean;
}

export interface ClassInfo {
  name: string;
  line: number;
  typeParameters: string[];
  extends: string[];
  implements: string[];
  members: number;
  isAbstract: boolean;
  isExported: boolean;
}

export interface EnumInfo {
  name: string;
  line: number;
  members: number;
  isConst: boolean;
  isExported: boolean;
}

export interface TypeInformation {
  typeAliases: TypeAlias[];
  interfaces: InterfaceInfo[];
  classes: ClassInfo[];
  enums: EnumInfo[];
  typeReferences: any[];
}

export interface FileAnalysisResult {
  filePath: string;
  entities: CodeEntity[];
  callExpressions: CallExpression[];
  typeInfo: TypeInformation;
  totalLines: number;
  sourceFile?: ts.SourceFile;
  language?: "rust" | "typescript" | "javascript";
}

export type EntityFilter = string | string[] | null;

export interface AnalysisOptions {
  context?: number;
  calls?: string | boolean;
  types?: boolean;
  showBuiltin?: boolean;
  callDepth?: number;
  flat?: boolean;
}